+++ flowtype/commands/__init__.py
@@ -2,10 +2,12 @@ from .add_pragma import FlowtypeAddPragma
 from .check_contents import FlowtypeCheckContents
 from .goto_definition import FlowtypeGotoDefinition
 from .view_type import FlowtypeViewType
+from .coverage import FlowtypeCoverage
 
 __all__ = [
     'FlowtypeAddPragma',
     'FlowtypeCheckContents',
     'FlowtypeGotoDefinition',
     'FlowtypeViewType',
+    'FlowtypeCoverage',
 ]