+++ flowtype/commands/coverage.py
@@ -6,31 +6,23 @@ from ..helpers import get_flow_bin, prepare_arguments
 logger = Logger()
 
 
-class FlowtypeViewType(BaseCommand):
-    """Run Flow type-at-pos and popup type definition."""
+class FlowtypeCoverage(BaseCommand):
+    """Run Flow coverage and highlight uncovered lines."""
 
     def get_cmd(self):
         """Construct cli command."""
         try:
             flow_bin = get_flow_bin()
         except ValueError as e:
-            logger.logger.error('type-at-pos %s' % e)
-            return
-
-        try:
-            project_root = self.get_project_root()
-        except ValueError as e:
-            logger.logger.error('type-at-pos %s' % e)
+            logger.logger.error('coverage %s' % e)
             return
 
         arguments = prepare_arguments(self.view)
 
         cmd = [
-            flow_bin, 'type-at-pos',
+            flow_bin, 'coverage',
             '--from', 'nuclide',
-            '--root', project_root,
-            '--path', arguments.file_name,
-            '--json', str(arguments.row + 1), str(arguments.col + 1)
+            '--json', arguments.file_name
         ]
 
         return cmd
@@ -41,19 +33,28 @@ class FlowtypeViewType(BaseCommand):
             error = error.decode('utf-8')
 
         if returncode != 0:
-            logger.logger.error('type-at-pos %s' % error)
+            logger.logger.error('coverage %s' % error)
             return
 
         logger.logger.debug(stdout)
 
         if stdout:
-            self.view.show_popup(stdout['type'])
+            expressions = stdout['expressions']
+            covered = expressions['covered_count']
+            uncovered = expressions['uncovered_count']
+            total = covered + uncovered
+            percentage = (covered * 100.0) / total
+            print(percentage)
+            self.view.set_status(
+                'flow_type',
+                'Flow: {}% coverage with {}/{} uncovered lines'
+                .format(round(percentage, 2), uncovered, covered))
         else:
             self.view.set_status('flow_type', 'Flow: no type found')
 
     def run(self, view):
-        """Execute `type-at-pos` command."""
-        logger.logger.debug('Running type-at-pos')
+        """Execute `coverage` command."""
+        logger.logger.debug('Running coverage')
 
         self.view.erase_status('flow_type')